@@ -1,33 +1,39 @@
 {
   "name": "@guessir/monorepo",
   "private": true,
+  "type": "module",
   "workspaces": [
     "packages/*"
   ],
   "scripts": {
     "build": "npm run build --workspaces --if-present",
     "clean": "npm run clean --workspaces --if-present",
-    "lint": "eslint packages/**/src/**/*.ts",
-    "start:ts:backend": "npm run start:ts -w @guessir/backend",
-    "start:ts:web": "npm run start:ts -w @guessir/web",
+    "lint": "eslint \"packages/**/src/**/*.ts\"",
+    "lint:fix": "npm run lint -- --fix",
+    "start:dev:backend": "npm run start:dev -w @guessir/backend",
+    "start:dev:web": "npm run start:dev -w @guessir/web",
+    "start:backend": "npm run start -w @guessir/backend",
     "test": "jest",
+    "test:clearCache": "jest --clearCache",
     "test:debug": "node --inspect-brk ./node_modules/jest/bin/jest.js --runInBand",
     "test:watch": "jest --watch"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.12.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
-    "jest": "^27.5.1",
+    "@swc/core": "^1.3.46",
+    "@swc/jest": "^0.2.24",
+    "@types/jest": "^29.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.57.1",
+    "@typescript-eslint/parser": "^5.57.1",
+    "del-cli": "^5.0.0",
+    "eslint": "^8.38.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "jest": "^29.5.0",
     "jest-sonar-reporter": "^2.0.0",
-    "prettier": "^2.6.2",
-    "ts-jest": "^27.1.4",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.3"
+    "prettier": "^2.8.7",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.4"
   },
   "jestSonar": {
     "reportPath": "./coverage"
