@@ -1,8 +1,8 @@
 import './styles.css';
 
-import { LexemeAnalyzer } from './lexemeAnalyzer';
-import { RenderController } from './renderers';
-import { hasTextParametersInUrl, loadText, parseTextIdFromUrl } from './utils';
+import { LexemeAnalyzer } from './lexemeAnalyzer/LexemeAnalyzer';
+import { RenderController } from './renderers/RenderController';
+import { hasTextParametersInUrl, loadText, parseTextIdFromUrl } from './utils/text';
 
 const defaultTitle = 'Hello!';
 
